+++ core/src/main/java/com/greenspacevoidnode/common/entity/world/WorldEntity.java
@@ -5,7 +5,7 @@ import com.greenspacevoidnode.common.entity.Entity;
 public class WorldEntity extends Entity {
 
 
-    public WorldEntity(String name, double x, double y, boolean isInvincible, boolean isTargetable, boolean canMove) {
+    public WorldEntity(String name, long x, long y, boolean isInvincible, boolean isTargetable, boolean canMove) {
         super(name, x, y, isInvincible, isTargetable, canMove);
     }
 }
